requirements.txt
Django==1.9.1
django-filter==0.12.0
djangorestframework==3.3.2
django-extensions==1.6.1
django-cors-headers==1.1.0
djangorestframework-jwt==1.7.2
